Finished extracting the traits of a request
This separates a request into being the sum of it's pieces. The request
now just holds a constructor, a fromGlobals method and a series of
traits
